+++ src/colors.ts
@@ -1,5 +1,22 @@
-// TODO:
-// 1) write documentation for color helpers.
-// 2) Write tests for color module
-// 3) Implement color module
-// 4) re-export color module from './index.ts'
+import Color from 'color';
+import * as colorNames from 'color-name';
+
+export type { default as ColorValue } from 'color';
+
+export type ColorName = keyof typeof colorNames;
+
+type ColorPlus = typeof Color & {
+  fromName(colorName: ColorName): Color;
+};
+
+export const color = Color as ColorPlus;
+
+export const rgb = color.rgb;
+export const hsl = color.hsl;
+
+/**
+ * Type-safe name to color mapper. Alias for `color(colorName)`
+ *
+ * 100% typing sugar, no substance.
+ */
+color.fromName = (colorName: ColorName) => color(colorName);